# Profile proto serialization with slice reused