--- a/api/buf.gen.yaml
+++ b/api/buf.gen.yaml
@@ -18,6 +18,11 @@ plugins:
       - features=marshal+unmarshal+size+pool+grpc+pool+clone
       - pool=github.com/grafana/pyroscope/api/gen/proto/go/google/v1.Profile
       - pool=github.com/grafana/pyroscope/api/gen/proto/go/google/v1.Sample
+      - pool=github.com/grafana/pyroscope/api/gen/proto/go/google/v1.Mapping
+      - pool=github.com/grafana/pyroscope/api/gen/proto/go/google/v1.Function
+      - pool=github.com/grafana/pyroscope/api/gen/proto/go/google/v1.Location
+      - pool=github.com/grafana/pyroscope/api/gen/proto/go/google/v1.Label
+      - pool=github.com/grafana/pyroscope/api/gen/proto/go/google/v1.Line
 
   - name: connect-go
     out: gen/proto/go