--- a/cmd/pyroscope/help.txt.tmpl
+++ b/cmd/pyroscope/help.txt.tmpl
@@ -1,4 +1,4 @@
-Usage of ./phlare:
+Usage of ./pyroscope:
   -api.base-url string
     	base URL for when the server is behind a reverse proxy with a different path
   -auth.multitenancy-enabled