… api/gen/proto/go/agent/v1/agent_vtproto.pb.go
… api/gen/proto/go/ingester/v1/ingester_vtproto.pb.go
… api/gen/proto/go/push/v1/push_vtproto.pb.go
… api/gen/proto/go/querier/v1/querier_vtproto.pb.go
… api/gen/proto/go/status/v1/status_vtproto.pb.go
… api/gen/proto/go/storegateway/v1/storegateway_vtproto.pb.go
… api/gen/proto/go/types/v1/types_vtproto.pb.go
… pkg/frontend/frontendpb/frontend_vtproto.pb.go
… pkg/og/convert/jfr/jfr_labels_vtproto.pb.go
… pkg/og/storage/tree/profile_vtproto.pb.go
… pkg/querier/stats/stats_vtproto.pb.go
… pkg/scheduler/schedulerpb/scheduler_vtproto.pb.go
… pkg/util/httpgrpc/httpgrpc_vtproto.pb.go