@kolesnikovae 2023-07-23
@korniltsev implemented a very efficient pprof parser with arenas and pools once – I'd consider resurrecting it