@cyriltovena 2023-07-24
Yeah that sounds good, the VT proto is a quick win though.