@cyriltovena 2023-07-24
Switched to a fork with the fix.

```
❯ benchstat before.txt after.txt
name             old time/op    new time/op    delta
FromRawBytes-16    8.07ms ± 2%    4.76ms ± 6%  -41.03%  (p=0.000 n=10+10)

name             old alloc/op   new alloc/op   delta
FromRawBytes-16    6.38MB ± 1%    0.21MB ±13%  -96.73%  (p=0.000 n=10+10)

name             old allocs/op  new allocs/op  delta
FromRawBytes-16     81.5k ± 0%      3.0k ± 0%  -96.29%  (p=0.000 n=10+8)
````